#51 Backend: Remove duplications, extract business logic to services

+++ src/main/java/com/example/restaurant_management_backend/services/ConfigService.java
@@ -10,8 +10,10 @@ import com.example.restaurant_management_backend.jpa.repositories.OpeningHourRep
 import com.example.restaurant_management_backend.mappers.ConfigMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 @RequiredArgsConstructor
@@ -25,35 +27,46 @@ public class ConfigService {
         return !configRepository.findAll().isEmpty();
     }
 
+    @Transactional
     public void initialize(ConfigAddCommand configAddCommand) {
         if (isSystemInitialized()) {
-            return;
+            throw new IllegalArgumentException("System został już zainicjalizowany");
         }
+
         deliveryPriceRepository.saveAll(configAddCommand.getDeliveryPricings());
         openingHourRepository.saveAll(configAddCommand.getOpeningHours());
         configRepository.save(configMapper.toConfig(configAddCommand));
     }
 
     public Config getConfig() {
-        List<Config> configs = configRepository.findAll();
-        if (configs.isEmpty()) {
-            return null;
-        }
-        return configs.getFirst();
+        return configRepository.findAll()
+                .stream()
+                .findFirst()
+                .orElseThrow(() -> new NoSuchElementException("System nie został zainicjalizowany"));
     }
 
     public List<DeliveryPricing> getDeliveryPrices() {
+        if (!isSystemInitialized()) {
+            throw new IllegalArgumentException("System nie został zainicjalizowany");
+        }
         return deliveryPriceRepository.findAll();
     }
 
     public List<OpeningHour> openingHours() {
+        if (!isSystemInitialized()) {
+            throw new IllegalArgumentException("System nie został zainicjalizowany");
+        }
         return openingHourRepository.findAll();
     }
 
+    @Transactional
     public void removeAll() {
+        if (!isSystemInitialized()) {
+            throw new IllegalArgumentException("Nie ma wgranej żadnej konfiguracji");
+        }
+
         deliveryPriceRepository.deleteAll();
         openingHourRepository.deleteAll();
-
         configRepository.deleteAll();
     }
 }
